fix: incorrect refactor file placement
