+++ packages/go/graphschema/graph.go
@@ -21,45 +21,7 @@ package graphschema
 
 import (
 	"github.com/specterops/bloodhound/dawgs/graph"
-	"github.com/specterops/bloodhound/graphschema/ad"
-	"github.com/specterops/bloodhound/graphschema/azure"
-	"github.com/specterops/bloodhound/graphschema/common"
 )
-
-var (
-	// Originates from BHE but copied here
-	meta         = graph.StringKind("Meta")
-	metaDetail   = graph.StringKind("MetaDetail")
-	metaIncludes = graph.StringKind("MetaIncludes")
-	metaKinds    = []graph.Kind{meta, metaDetail, metaIncludes}
-
-	unknownKind = graph.StringKind("Unknown")
-
-	// Used for quick O(1) kind lookups
-	ValidKinds = buildValidKinds()
-)
-
-func buildValidKinds() map[graph.Kind]bool {
-	var (
-		validKinds = make(map[graph.Kind]bool)
-		kindSlices = []graph.Kinds{
-			ad.NodeKinds(),
-			ad.Relationships(),
-			azure.NodeKinds(),
-			azure.Relationships(),
-			common.NodeKinds(),
-			common.Relationships(),
-		}
-	)
-
-	for _, kindSlice := range kindSlices {
-		for _, kind := range kindSlice {
-			validKinds[kind] = true
-		}
-	}
-
-	return validKinds
-}
 
 type KindDescriptor struct {
 	Kind graph.Kind
@@ -77,33 +39,4 @@ type Path struct {
 	Outbound      KindDescriptor
 	Inbound       KindDescriptor
 	Relationships []KindDescriptor
-}
-
-func PrimaryNodeKind(kinds graph.Kinds) graph.Kind {
-	var (
-		resultKind = unknownKind
-		baseKind   = resultKind
-	)
-
-	for _, kind := range kinds {
-		// If this is a BHE meta kind, return early
-		if kind.Is(metaKinds...) {
-			return meta
-		} else if kind.Is(ad.Entity, azure.Entity) {
-			baseKind = kind
-		} else if kind.Is(ad.LocalGroup) {
-			// Allow ad.LocalGroup to overwrite NodeKindUnknown, but nothing else
-			if resultKind == unknownKind {
-				resultKind = kind
-			}
-		} else if ValidKinds[kind] {
-			resultKind = kind
-		}
-	}
-
-	if resultKind.Is(unknownKind) {
-		return baseKind
-	} else {
-		return resultKind
-	}
 }
